Update mkinit
- Change /usr/lib/plan9 prefix to /opt/plan9
- Update bitlbee rule for bitlbee 3.0
- Cleanup spamassassin rule

+++ src/initctld.c
@@ -132,12 +132,12 @@ void process_request(struct init_request *request)
 		break;
 	case INIT_CMD_SETENV:
 		printf("eval export %.*s\n",
-				sizeof(request->i.data),
+				(int)sizeof(request->i.data),
 				request->i.data);
 		break;
 	case INIT_CMD_CHANGECONS:
 		printf("changeconsole %.*s\n",
-				sizeof(request->i.bsd.reserved),
+				(int)sizeof(request->i.bsd.reserved),
 				request->i.bsd.reserved);
 		break;
 	default:
